[GC] Make the interaction between GC handles and reference queues well defined.
The previous usage of those two features did not intersect but it now does with correct handling of Context objects.

Thus we set the relationship between the two to be the following:

- At the time a reference queue entry is processed, all GC handles that did point to
the same object will have been cleared.

- During unload and cleanup, GC handles are cleared first, meaning they will hold no values
when reference queue entries are processed.

- Reference queue callbacks should not rely of the values of GC handles, as they might not
hold the expected values due to unload and cleanup.