[DONT MERGE] This is a dump of code to handle the frequent crash on JI freeing code that has plaguing XS.
The fix is the next_puppy_to_free linked list.

Here's the explanation:

What we know:
In mono_jit_info_table_free, we double free in line 139 (freeing the tombstone)

Theory:
What we're actually freeing is not the tombstone, but an actual JI that was freed from jit_info_free_queue.

For that to happen we need the following to happen:

When the JI was freed:
	num_jit_info_tables > 1
	The tombstone was installed on a chunk only in used by the new table.

For a chunk to only exist on the new table:
	We overflow the chunk or
	We purify it because there are too many tombstones

Finally, after we free the JI, for some reason memory is reused and it turns to zero.

Sequence of events:

- Add JI Y to chunk X
- Keep adding JIs to chunk X until it overflows
at this point num_jit_info_tables == 2
at this point the new table doesn't include chunk X
- Remove JI Y
JI is tombstone'd only on the new chunk and not on X
- Collect hazard pointers
boom!

Event sequence - using (?) to denote thread:

(1) Allocs a fnptr for delegate D, that's our JI Y
(1) Allocs a ton more until it overflows, now Y is on both chunks
(1) Make D unreachable
(1) Call GC.Collect
(1) Restart the world and kick the Finalizer thread
(2) Finalizer thread goes on and call mono_jit_info_table_remove (Y)
	-> Since mono_thread_hazardous_try_free_some has not run, num_jit_info_tables == 2
(1) GC ends, call mono_thread_hazardous_try_free_some
	-> boom