[offsets-tool] Add a couple of parameters (verbose, platform) and fix usage outside of XI.
As part of making this tool usable outside of XI, I had to make a few adjustments.

First add a platform argument that can be used instead of --monodroid or --maccore.
That allows the tools to work without using the hardcoded structure of those repos.

Second, I made the iOS setup code respect the --target dir option instead of only using
the hardcoded one from XI.