[wasm] Skeleton for wasm specific implementations of a bunch of OS stuff.
This commit is incomplete because we need to implement / investigate all missing functions that I added as comments in w32process-wasm.c.
Savings for a release build are of the order of 100kb for interp-release.

The icalls should be all implemented as throwing PSNE.

The non-icall functions require some investigation.